+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 ### Added
 - Send unscheduled alarm event
 
+### Changed
+- Improve UI (add edit alarm)
+
+### Fixed
+- Fix issue during alarm scheduling
+
 ### Fixed
 - Fix issue when multiple alarms scheduled for today and tomorrow. Only one was scheduled